Separating Polygonizer
I've split the polygonizer into a separate class.

There is a base polygonizer class and a specific marching cubes class.

The constructor defines what scene is polygonized and how many voxels the
longest side will have. i.e, if I define the maximum voxels to be 5, and the
bounding volume is a cube, I will have a total of 125 voxels calculated.

The new construction will make it easier for implementing a multi-threaded
architecture.