Adding new mesh to polygonizer
I've added the newer mesh class to the polygonizer.
There appears to be an issue with the meaning of the indices.
Those will be fixed shortly.

TODO: The indices stores in the INDEX triangle will be for the position in the
vectors. That is, we will have to add one to the index when exporting it, and
subtract one from the index when importing it.