Changed Index Meaning
For the purpose of possibly storing uv coordinates in the future, I've decided
to store the index as it shows up in an obj file. That is, with an offset of
one. An index of 0 is invalid, so it should not be referenced. An index of 1
will dereference index 0 in the arrays. Why not use -1 as an invalid reference?
I want to use unsigned int and feel that simply adding the ability to use -1 as
an invalid index is worth losing half the possible numbers that 32 bits can
represent.

So:
0 is invalid.
1 -> m_verts[0]
2 -> m_verts[1]
...

Why is this important:
My objects are not uv-unwrapped, so the likely hood that an object is going to
have uv coordinates is low. Furthermore, we have no way of creating a good
projection (we could project onto a coordinate plane, but that is not
necessarily good), so I'm allowing this to be blank. For the purpose of
consistency, it is good to make sure that the indices for vertices, normals,
and uv coordinates have the same semantic meaning.
A helper function an be created that subtracts one from the index, but it
should not be necessary.